mosquitto_MQTT/requirements.txt
# MQTT 클라이언트 라이브러리
paho-mqtt==1.6.1

# 시간 관련 유틸리티
schedule==1.2.0

# 데이터 처리 및 분석
numpy==1.24.3
pandas==2.0.3

# JSON 및 YAML 처리 (내장 모듈이지만 명시적 표기)
# json - 내장 모듈
# yaml - PyYAML로 대체 가능하지만 선택적

# 날짜/시간 처리 (내장 모듈)
# datetime - 내장 모듈

# 랜덤 데이터 생성 (내장 모듈) 
# random - 내장 모듈

# 로깅 (내장 모듈)
# logging - 내장 모듈

# 설정 파일 처리 (선택적)
PyYAML==6.0.1

# 프로세스 관리 및 신호 처리 (내장 모듈)
# signal - 내장 모듈
# threading - 내장 모듈

# 시스템 유틸리티 (내장 모듈)
# time - 내장 모듈
# os - 내장 모듈
# sys - 내장 모듈

# 추가 유틸리티 (선택적)
colorama==0.4.6  # 터미널 색상 출력
tqdm==4.65.0     # 진행률 표시바 (선택적)

# 개발 도구 (선택적)
python-dotenv==1.0.0  # 환경변수 관리